refactor: split src/util.rs into focused sub-modules
Replace the flat src/util.rs with src/util/ directory containing:
- formatting.rs: string truncation helpers
- validation.rs: input validation (branch names, slugs, env vars, gh args)

Re-exports in mod.rs preserve all existing call sites unchanged.

Closes #362
